+++ packages/bridge/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.13](https://github.com/rowthan/pagenote/compare/@pagenote/bridge@1.0.8...@pagenote/bridge@1.0.13) (2024-01-15)
+
+
+### Features
+
+* broadcast API for extension ([aa9e1a0](https://github.com/rowthan/pagenote/commit/aa9e1a029ec85e1f0c7c318349bf024ee8e5ece4))
+* catch ondisconnect on request ([9eec683](https://github.com/rowthan/pagenote/commit/9eec68348dc59a562eeafeb7254034608127f556))
+
+
+
+
+
+## [1.0.12](https://github.com/rowthan/pagenote/compare/@pagenote/bridge@1.0.11...@pagenote/bridge@1.0.12) (2024-01-15)
+
+**Note:** Version bump only for package @pagenote/bridge
+
+
+
+
+
+## [1.0.11](https://github.com/rowthan/pagenote/compare/@pagenote/bridge@1.0.10...@pagenote/bridge@1.0.11) (2024-01-15)
+
+**Note:** Version bump only for package @pagenote/bridge
+
+
+
+
+
+## [1.0.10](https://github.com/rowthan/pagenote/compare/@pagenote/bridge@1.0.9...@pagenote/bridge@1.0.10) (2024-01-13)
+
+
+### Features
+
+* catch ondisconnect on request ([9b1b3ba](https://github.com/rowthan/pagenote/commit/9b1b3baf0ac0f0465d66e6018d2b1ed023ec0b74))
+
+
+
+
+
+## [1.0.9](https://github.com/rowthan/pagenote/compare/@pagenote/bridge@1.0.8...@pagenote/bridge@1.0.9) (2024-01-13)
+
+
+### Features
+
+* broadcast API for extension ([4732dd8](https://github.com/rowthan/pagenote/commit/4732dd8d53546a19a6a32c0041d5ceea6145e976))
+
+
+
+
+
 ## [1.0.8](https://github.com/rowthan/pagenote/compare/@pagenote/bridge@1.0.3...@pagenote/bridge@1.0.8) (2024-01-07)
 
 **Note:** Version bump only for package @pagenote/bridge